Final Push - Ready for Review
Added four arrays with item information for 4 items, including price, description, and quantity. Added a for loop that dynamically updates the item's information based on how many times the for loop has gone through the loop. Added code to change the webpage to include all the information, as well as correctly price that information, so that it updates when the arrays do. Changed file names and added author and date.